Fix for loss of GLTF override data due to objects eviction from the rendered scene
Commit 64c541f40191dee0fb6b3b0c09d0d9bfacb454a5 introduced a bug causing the indue removal
of GLTF override data whenever an object is 'killed' when not rendered any
more (e.g. an object in a connected neighbour region finding itself out of
draw distance when the user moves or cams around).

Since this data is not resent by the sim server until the next reconnection
of the said sim, the object may re-rez later with missing overrides.

This commit simply comments out the culprit line.